Refactor: Changes location of commonly used objects in tests.
- Removed `TestsData.kt` from common-feature module.
- Now, if need be, every module has internally its own reusable Constants used for its tests.
- If something is used across more than one modules, then it resides in the test-feature module.
