+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/ui/documents/detail/model/DocumentDetailsUi.kt
@@ -14,22 +14,20 @@
  * governing permissions and limitations under the Licence.
  */
 
-package eu.europa.ec.commonfeature.model
+package eu.europa.ec.dashboardfeature.ui.documents.detail.model
 
-data class DocumentCredentialsInfoUi(
-    val availableCredentials: Int,
-    val totalCredentials: Int,
-    val title: String,
-    val collapsedInfo: CollapsedInfo? = null,
-    val expandedInfo: ExpandedInfo? = null,
-) {
-    data class CollapsedInfo(
-        val moreInfoText: String,
-    )
+import eu.europa.ec.corelogic.model.DocumentIdentifier
+import eu.europa.ec.eudi.wallet.document.DocumentId
+import eu.europa.ec.uilogic.component.wrap.ExpandableListItem
 
-    data class ExpandedInfo(
-        val subtitle: String,
-        val updateNowButtonText: String?,
-        val hideButtonText: String,
-    )
+enum class DocumentIssuanceStateUi {
+    Issued, Pending, Failed, Expired, Revoked
 }
+
+data class DocumentDetailsUi(
+    val documentId: DocumentId,
+    val documentName: String,
+    val documentIdentifier: DocumentIdentifier,
+    val documentIssuanceStateUi: DocumentIssuanceStateUi,
+    val documentClaims: List<ExpandableListItem>,
+)
