+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/ui/documents/list/DocumentsScreen.kt
@@ -59,11 +59,11 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
-import eu.europa.ec.commonfeature.model.DocumentUiIssuanceState
 import eu.europa.ec.corelogic.model.DocumentCategory
 import eu.europa.ec.corelogic.model.DocumentIdentifier
 import eu.europa.ec.corelogic.util.CoreActions
 import eu.europa.ec.dashboardfeature.model.SearchItemUi
+import eu.europa.ec.dashboardfeature.ui.documents.detail.model.DocumentIssuanceStateUi
 import eu.europa.ec.dashboardfeature.ui.documents.list.model.DocumentUi
 import eu.europa.ec.resourceslogic.R
 import eu.europa.ec.resourceslogic.theme.values.warning
@@ -368,8 +368,8 @@ private fun DocumentCategory(
                 item = documentItem.uiData,
                 onItemClick = {
                     val onItemClickEvent = if (
-                        documentItem.documentIssuanceState == DocumentUiIssuanceState.Pending
-                        || documentItem.documentIssuanceState == DocumentUiIssuanceState.Failed
+                        documentItem.documentIssuanceState == DocumentIssuanceStateUi.Pending
+                        || documentItem.documentIssuanceState == DocumentIssuanceStateUi.Failed
                     ) {
                         Event.BottomSheet.DeferredDocument.DeferredNotReadyYet.DocumentSelected(
                             documentId = documentItem.uiData.itemId
@@ -380,11 +380,11 @@ private fun DocumentCategory(
                     onEventSend(onItemClickEvent)
                 },
                 supportingTextColor = when (documentItem.documentIssuanceState) {
-                    DocumentUiIssuanceState.Issued -> null
-                    DocumentUiIssuanceState.Pending -> MaterialTheme.colorScheme.warning
-                    DocumentUiIssuanceState.Failed -> MaterialTheme.colorScheme.error
-                    DocumentUiIssuanceState.Expired -> MaterialTheme.colorScheme.error
-                    DocumentUiIssuanceState.Revoked -> MaterialTheme.colorScheme.error
+                    DocumentIssuanceStateUi.Issued -> null
+                    DocumentIssuanceStateUi.Pending -> MaterialTheme.colorScheme.warning
+                    DocumentIssuanceStateUi.Failed -> MaterialTheme.colorScheme.error
+                    DocumentIssuanceStateUi.Expired -> MaterialTheme.colorScheme.error
+                    DocumentIssuanceStateUi.Revoked -> MaterialTheme.colorScheme.error
                 }
             )
         }
@@ -594,7 +594,7 @@ private fun DocumentsScreenPreview() {
             val validUntil = "Valid Until"
             val documentsList = listOf(
                 DocumentUi(
-                    documentIssuanceState = DocumentUiIssuanceState.Issued,
+                    documentIssuanceState = DocumentIssuanceStateUi.Issued,
                     uiData = ListItemData(
                         itemId = "id1",
                         mainContentData = ListItemMainContentData.Text(text = "Document 1"),
@@ -607,7 +607,7 @@ private fun DocumentsScreenPreview() {
                     documentCategory = DocumentCategory.Government
                 ),
                 DocumentUi(
-                    documentIssuanceState = DocumentUiIssuanceState.Issued,
+                    documentIssuanceState = DocumentIssuanceStateUi.Issued,
                     uiData = ListItemData(
                         itemId = "id2",
                         mainContentData = ListItemMainContentData.Text(text = "Document 2"),
@@ -620,7 +620,7 @@ private fun DocumentsScreenPreview() {
                     documentCategory = DocumentCategory.Government
                 ),
                 DocumentUi(
-                    documentIssuanceState = DocumentUiIssuanceState.Issued,
+                    documentIssuanceState = DocumentIssuanceStateUi.Issued,
                     uiData = ListItemData(
                         itemId = "id3",
                         mainContentData = ListItemMainContentData.Text(text = "Document 3"),
@@ -633,7 +633,7 @@ private fun DocumentsScreenPreview() {
                     documentCategory = DocumentCategory.Finance
                 ),
                 DocumentUi(
-                    documentIssuanceState = DocumentUiIssuanceState.Issued,
+                    documentIssuanceState = DocumentIssuanceStateUi.Issued,
                     uiData = ListItemData(
                         itemId = "id4",
                         mainContentData = ListItemMainContentData.Text(text = "Document 4"),
