+++ resources-logic/src/main/res/values/strings.xml
@@ -179,7 +179,6 @@
     <string name="request_gender_female">Female</string>
     <string name="request_gender_not_known">Not known</string>
     <string name="request_gender_not_applicable">Not applicable</string>
-    <string name="request_element_identifier_not_available">Not available</string>
 
     <!-- Loading Screen -->
     <string name="loading_header_description">Please wait…</string>
