--- test-feature/src/main/java/eu/europa/ec/testfeature/Constants.kt
@@ -1,690 +1,0 @@
-/*
- * Copyright (c) 2023 European Commission
- *
- * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the European
- * Commission - subsequent versions of the EUPL (the "Licence"); You may not use this work
- * except in compliance with the Licence.
- *
- * You may obtain a copy of the Licence at:
- * https://joinup.ec.europa.eu/software/page/eupl
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the Licence is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF
- * ANY KIND, either express or implied. See the Licence for the specific language
- * governing permissions and limitations under the Licence.
- */
-
-package eu.europa.ec.testfeature
-
-import eu.europa.ec.eudi.wallet.document.IssuedDocument
-import eu.europa.ec.eudi.wallet.document.format.DocumentData
-import eu.europa.ec.eudi.wallet.document.format.DocumentFormat
-import eu.europa.ec.eudi.wallet.document.format.MsoMdocData
-import eu.europa.ec.eudi.wallet.document.format.MsoMdocFormat
-import eu.europa.ec.eudi.wallet.document.format.SdJwtVcData
-import eu.europa.ec.eudi.wallet.document.format.SdJwtVcFormat
-import eu.europa.ec.eudi.wallet.document.metadata.IssuerMetadata
-import org.mockito.Mockito.mock
-import org.mockito.kotlin.whenever
-import org.multipaz.document.NameSpacedData
-import java.net.URI
-import java.time.Instant
-import java.util.Locale
-
-const val mockedGenericErrorMessage = "resourceProvider's genericErrorMessage"
-const val mockedPlainFailureMessage = "failure message"
-
-val mockedExceptionWithMessage = RuntimeException("Exception to test interactor.")
-val mockedExceptionWithNoMessage = RuntimeException()
-
-val mockedDefaultLocale: Locale = Locale.ENGLISH
-
-const val mockedChangeLogUrl = "https://example.com/changelog"
-const val mockedOldestDocumentCreationDate = "2000-01-25T14:25:00.073Z"
-const val mockedDocumentCreationDate = "2024-01-25T14:25:00.073Z"
-const val mockedDocumentValidUntilDate = "2030-05-13T14:25:00.073Z"
-const val mockedOldestPidId = "000000"
-const val mockedPidId = "000001"
-const val mockedMdlId = "000002"
-const val mockedSdJwtPidId = "000003"
-const val mockedPidDocName = "EU PID"
-const val mockedMdlDocName = "mDL"
-const val mockedDocumentAvailableCredentials = 3
-const val mockedDocumentTotalCredentials = 15
-const val mockedBookmarkId = "mockedBookmarkId"
-const val mockedVerifierIsTrusted = true
-const val mockedNotifyOnAuthenticationFailure = false
-val mockedPidFields: Map<String, ByteArray> = mapOf(
-    "family_name" to byteArrayOf(105, 65, 78, 68, 69, 82, 83, 83, 79, 78),
-    "given_name" to byteArrayOf(99, 74, 65, 78),
-    "birth_date" to byteArrayOf(-39, 3, -20, 106, 49, 57, 56, 53, 45, 48, 51, 45, 51, 48),
-    "age_over_18" to byteArrayOf(-11),
-    "age_over_15" to byteArrayOf(-11),
-    "age_over_21" to byteArrayOf(-11),
-    "age_over_60" to byteArrayOf(-12),
-    "age_over_65" to byteArrayOf(-12),
-    "age_over_68" to byteArrayOf(-12),
-    "age_in_years" to byteArrayOf(24, 38),
-    "age_birth_year" to byteArrayOf(25, 7, -63),
-    "family_name_birth" to byteArrayOf(105, 65, 78, 68, 69, 82, 83, 83, 79, 78),
-    "given_name_birth" to byteArrayOf(99, 74, 65, 78),
-    "birth_place" to byteArrayOf(102, 83, 87, 69, 68, 69, 78),
-    "birth_country" to byteArrayOf(98, 83, 69),
-    "birth_state" to byteArrayOf(98, 83, 69),
-    "birth_city" to byteArrayOf(107, 75, 65, 84, 82, 73, 78, 69, 72, 79, 76, 77),
-    "resident_address" to byteArrayOf(
-        111,
-        70,
-        79,
-        82,
-        84,
-        85,
-        78,
-        65,
-        71,
-        65,
-        84,
-        65,
-        78,
-        32,
-        49,
-        53
-    ),
-    "resident_country" to byteArrayOf(98, 83, 69),
-    "resident_state" to byteArrayOf(98, 83, 69),
-    "resident_city" to byteArrayOf(107, 75, 65, 84, 82, 73, 78, 69, 72, 79, 76, 77),
-    "resident_postal_code" to byteArrayOf(101, 54, 52, 49, 51, 51),
-    "resident_street" to byteArrayOf(108, 70, 79, 82, 84, 85, 78, 65, 71, 65, 84, 65, 78),
-    "resident_house_number" to byteArrayOf(98, 49, 50),
-    "gender" to byteArrayOf(1),
-    "nationality" to byteArrayOf(98, 83, 69),
-    "issuance_date" to byteArrayOf(
-        -64,
-        116,
-        50,
-        48,
-        48,
-        57,
-        45,
-        48,
-        49,
-        45,
-        48,
-        49,
-        84,
-        48,
-        48,
-        58,
-        48,
-        48,
-        58,
-        48,
-        48,
-        90
-    ),
-    "expiry_date" to byteArrayOf(
-        -64,
-        116,
-        50,
-        48,
-        53,
-        48,
-        45,
-        48,
-        51,
-        45,
-        51,
-        48,
-        84,
-        48,
-        48,
-        58,
-        48,
-        48,
-        58,
-        48,
-        48,
-        90
-    ),
-    "issuing_authority" to byteArrayOf(99, 85, 84, 79),
-    "document_number" to byteArrayOf(105, 49, 49, 49, 49, 49, 49, 49, 49, 52),
-    "administrative_number" to byteArrayOf(106, 57, 48, 49, 48, 49, 54, 55, 52, 54, 52),
-    "issuing_country" to byteArrayOf(98, 83, 69),
-    "issuing_jurisdiction" to byteArrayOf(100, 83, 69, 45, 73),
-)
-val mockedMdlFields: Map<String, ByteArray> = mapOf(
-    "family_name" to byteArrayOf(105, 65, 78, 68, 69, 82, 83, 83, 79, 78),
-    "given_name" to byteArrayOf(99, 74, 65, 78),
-    "birth_date" to byteArrayOf(-39, 3, -20, 106, 49, 57, 56, 53, 45, 48, 51, 45, 51, 48),
-    "issue_date" to byteArrayOf(
-        -64,
-        116,
-        50,
-        48,
-        48,
-        57,
-        45,
-        48,
-        49,
-        45,
-        48,
-        49,
-        84,
-        48,
-        48,
-        58,
-        48,
-        48,
-        58,
-        48,
-        48,
-        90
-    ),
-    "expiry_date" to byteArrayOf(
-        -64,
-        116,
-        50,
-        48,
-        53,
-        48,
-        45,
-        48,
-        51,
-        45,
-        51,
-        48,
-        84,
-        48,
-        48,
-        58,
-        48,
-        48,
-        58,
-        48,
-        48,
-        90
-    ),
-    "issuing_country" to byteArrayOf(98, 83, 69),
-    "issuing_authority" to byteArrayOf(99, 85, 84, 79),
-    "document_number" to byteArrayOf(105, 49, 49, 49, 49, 49, 49, 49, 49, 52),
-    "portrait" to byteArrayOf(98, 83, 69),
-    "un_distinguishing_sign" to byteArrayOf(97, 83),
-    "administrative_number" to byteArrayOf(106, 57, 48, 49, 48, 49, 54, 55, 52, 54, 52),
-    "sex" to byteArrayOf(1),
-    "height" to byteArrayOf(24, -76),
-    "weight" to byteArrayOf(24, 91),
-    "eye_colour" to byteArrayOf(101, 98, 108, 97, 99, 107),
-    "hair_colour" to byteArrayOf(101, 98, 108, 97, 99, 107),
-    "birth_place" to byteArrayOf(102, 83, 87, 69, 68, 69, 78),
-    "resident_address" to byteArrayOf(
-        111,
-        70,
-        79,
-        82,
-        84,
-        85,
-        78,
-        65,
-        71,
-        65,
-        84,
-        65,
-        78,
-        32,
-        49,
-        53
-    ),
-    "portrait_capture_date" to byteArrayOf(
-        -64,
-        116,
-        50,
-        48,
-        50,
-        51,
-        45,
-        48,
-        51,
-        45,
-        50,
-        51,
-        84,
-        48,
-        48,
-        58,
-        48,
-        48,
-        58,
-        48,
-        48,
-        90
-    ),
-    "signature_usual_mark" to byteArrayOf(98, 83, 69),
-    "age_in_years" to byteArrayOf(24, 38),
-    "age_birth_year" to byteArrayOf(25, 7, -63),
-    "issuing_jurisdiction" to byteArrayOf(100, 83, 69, 45, 73),
-    "nationality" to byteArrayOf(98, 83, 69),
-    "resident_city" to byteArrayOf(102, 83, 87, 69, 68, 69, 78),
-    "resident_state" to byteArrayOf(98, 83, 69),
-    "resident_postal_code" to byteArrayOf(101, 54, 52, 49, 51, 51),
-    "resident_country" to byteArrayOf(98, 83, 69),
-    "family_name_national_character" to byteArrayOf(
-        105,
-        65,
-        78,
-        68,
-        69,
-        82,
-        83,
-        83,
-        79,
-        78
-    ),
-    "given_name_national_character" to byteArrayOf(99, 74, 65, 78),
-    "age_over_15" to byteArrayOf(-11),
-    "age_over_18" to byteArrayOf(-11),
-    "age_over_21" to byteArrayOf(-11),
-    "age_over_60" to byteArrayOf(-12),
-    "age_over_65" to byteArrayOf(-12),
-    "age_over_68" to byteArrayOf(-12),
-)
-
-const val mockedSdJwtFullPidFields: String =
-    "eyJhbGciOiAiRVMyNTYiLCAidHlwIjogImRjK3NkLWp3dCIsICJ4NWMiOiBbIk1JSUMzekNDQW9XZ0F3SUJBZ0lVZjNsb2hUbURNQW1TL1lYL3E0aHFvUnlKQjU0d0NnWUlLb1pJemowRUF3SXdYREVlTUJ3R0ExVUVBd3dWVUVsRUlFbHpjM1ZsY2lCRFFTQXRJRlZVSURBeU1TMHdLd1lEVlFRS0RDUkZWVVJKSUZkaGJHeGxkQ0JTWldabGNtVnVZMlVnU1cxd2JHVnRaVzUwWVhScGIyNHhDekFKQmdOVkJBWVRBbFZVTUI0WERUSTFNRFF4TURFME16YzFNbG9YRFRJMk1EY3dOREUwTXpjMU1Wb3dVakVVTUJJR0ExVUVBd3dMVUVsRUlFUlRJQzBnTURFeExUQXJCZ05WQkFvTUpFVlZSRWtnVjJGc2JHVjBJRkpsWm1WeVpXNWpaU0JKYlhCc1pXMWxiblJoZEdsdmJqRUxNQWtHQTFVRUJoTUNWVlF3V1RBVEJnY3Foa2pPUFFJQkJnZ3Foa2pPUFFNQkJ3TkNBQVM3V0FBV3FQemUwVXMzejhwYWp5VlBXQlJtclJiQ2k1WDJzOUd2bHliUXl0d1R1bWNabmVqOUJrTGZBZ2xsb1g1dHYrTmdXZkRmZ3QvMDZzKzV0VjRsbzRJQkxUQ0NBU2t3SHdZRFZSMGpCQmd3Rm9BVVlzZVVSeWk5RDZJV0lLZWF3a21VUlBFQjA4Y3dHd1lEVlIwUkJCUXdFb0lRYVhOemRXVnlMbVYxWkdsM0xtUmxkakFXQmdOVkhTVUJBZjhFRERBS0JnZ3JnUUlDQUFBQkFqQkRCZ05WSFI4RVBEQTZNRGlnTnFBMGhqSm9kSFJ3Y3pvdkwzQnlaWEJ5YjJRdWNHdHBMbVYxWkdsM0xtUmxkaTlqY213dmNHbGtYME5CWDFWVVh6QXlMbU55YkRBZEJnTlZIUTRFRmdRVXFsL29weGtRbFl5MGxsYVRvUGJERS9teUVjRXdEZ1lEVlIwUEFRSC9CQVFEQWdlQU1GMEdBMVVkRWdSV01GU0dVbWgwZEhCek9pOHZaMmwwYUhWaUxtTnZiUzlsZFMxa2FXZHBkR0ZzTFdsa1pXNTBhWFI1TFhkaGJHeGxkQzloY21Ob2FYUmxZM1IxY21VdFlXNWtMWEpsWm1WeVpXNWpaUzFtY21GdFpYZHZjbXN3Q2dZSUtvWkl6ajBFQXdJRFNBQXdSUUloQU5KVlNEc3FUM0lrR2NLV1dnU2V1YmtET2RpNS9VRTliMUdGL1g1ZlFSRmFBaUJwNXQ2dEhoOFh3RmhQc3R6T0hNb3B2QkQvR3dtczBSQVVnbVNuNmt1OEdnPT0iXX0.eyJfc2QiOiBbIjJZeUtSV2U3NEQzMjJVRkFHemRMaEhUXzM3eGtaYUd4UmQxMU5oc0ZZQ1EiLCAiN0ZuQi1UOXkyWEoyLVJ2bW85NzVXTWlfYUhoTUZiU1ZJWDI5YkNIa05hZyIsICI3VndBOWI3U1JONlJjaTRnNnBUTFlNU3pyY25wWkRXXzk4ZWFDQXlJanFnIiwgIjl3QWhoeEhGUjRtbGdjOWphNHlqSWlyNTVBYVVhYXR6bzc0eEh0UnBrZEkiLCAiRkNfNEdpNV83MXVVODFmLXA3UEFxYXpaOUtoY3hiMkNVclU0X29pSGs1QSIsICJMdnJpdnRmOUpzQXFzaTUwRUpEQlUtcGpacXVkVkZ6VE93RUtzT2pHQlNFIiwgIlJfaTRqNlRLcklmRzBnY2dqOHoxYTQ1T0RyZFlkZTJlQnBvZjFEN2paSUkiLCAiYThsMHhOY0lTaG1EV3RtSkVMMm80ZHJib2YzbzdZV3BRRHg0UTNPQWFWNCIsICJidVo4N0FUQU81eXdmRzM2T1ZfMEl5NEdUZ2JWSVdqOE13ajkwWFFVVGtNIiwgImNGd3NvaFcwRktvWmlpQW5oUEtrU0hBdFJoSGlGNE15LU5jTGJuNFMtZGciLCAiY2xQcGFqWnVuWWZKOE9XdnZVVkVtb0lDREJYZHRLMk9YZGpQdWdrTGdmTSIsICJoQVlQVzhNQXlDYkg3dGhJN2p6UFFEMXpVMExrWTVvVUpHY1l5d3dPTVlvIiwgImkzWHlXYmlXY2dxcFBlWFpZZDBYXzFUUi00YmJETVUtQjVJZDZrMGtJWlUiLCAibjlJQ0lPdF9UeHNlRGpvNHhCSjgxWVZlUkVqZGYwMzN6Vnk0U1N0dlJlVSIsICJvTzd3c1lxT0VaYk9RWVNQTzZXX3lRdkNOaE5RWDZQcS1CUl9IRm5OR1drIiwgInhkMEhRYVRlWHIwcERyb1pudjJWeTFlTUlkSTNDZmhUOHFJOGFRMzFBQ1EiXSwgImlzcyI6ICJodHRwczovL2lzc3Vlci5ldWRpdy5kZXYiLCAiaWF0IjogMTc0Nzg2ODQwMCwgImV4cCI6IDE3NTU2NDQ0MDAsICJ2Y3QiOiAidXJuOmV1LmV1cm9wYS5lYy5ldWRpOnBpZDoxIiwgInN0YXR1cyI6IHsiaWRlbnRpZmllcl9saXN0IjogeyJpZCI6ICIzNzc4IiwgInVyaSI6ICJodHRwczovL2lzc3Vlci5ldWRpdy5kZXYvaWRlbnRpZmllcl9saXN0L0ZDL2V1LmV1cm9wYS5lYy5ldWRpLnBpZC4xLzk5ODdkY2E1LWYzMDEtNGY3NS1hOWVmLWQ2MmIwOTVhMGVmNCJ9LCAic3RhdHVzX2xpc3QiOiB7ImlkeCI6IDM3NzgsICJ1cmkiOiAiaHR0cHM6Ly9pc3N1ZXIuZXVkaXcuZGV2L3Rva2VuX3N0YXR1c19saXN0L0ZDL2V1LmV1cm9wYS5lYy5ldWRpLnBpZC4xLzk5ODdkY2E1LWYzMDEtNGY3NS1hOWVmLWQ2MmIwOTVhMGVmNCJ9fSwgIl9zZF9hbGciOiAic2hhLTI1NiIsICJjbmYiOiB7Imp3ayI6IHsia3R5IjogIkVDIiwgImNydiI6ICJQLTI1NiIsICJ4IjogImxfeXNWcFNCOGVaekxRRURVdTMxemxXVWRwdnU2R0RuNjJPaTh0OTh5ZjQiLCAieSI6ICJhaC1ZdUlQVFAtWmFLSzJhc21YWGlXaHFJNEZQTVRDZk5kVzNyOENFNVRZIn19fQ.xU58emwc4_r5w5uQ42foWskLtrlHNLNKMb--ZZAKhDBB57c9XYIRQbk9hxNEo7TyN9lem_muzAQX4h8JqFBR9Q~WyJaNFFkTlBaOTVVVnQ0UlJkemYxUG9nIiwgImZhbWlseV9uYW1lIiwgIkFOREVSU1NPTiJd~WyJybXY5UlNmOGY2X280dFgtYnhhclF3IiwgImdpdmVuX25hbWUiLCAiSkFOIl0~WyJIXzlZcmhlU19wRW1nbl8tWTliWE53IiwgImJpcnRoX2RhdGUiLCAiMTk4NS0wMy0zMCJd~WyI0a2xZUzgxWE5YV21TU3pYUlRQUld3IiwgImxvY2FsaXR5IiwgIktBVFJJTkVIT0xNIl0~WyJvMTRlVFE1MTBvN3dvX0NZeXBrUnhRIiwgInJlZ2lvbiIsICJTRSJd~WyJhVU9JUHV0UkFUTk5OTjAxejJsVndnIiwgImNvdW50cnkiLCAiU1dFREVOIl0~WyJ1ZXk3NVdaQmFSTDZZNkIyS18zbFpRIiwgInBsYWNlX29mX2JpcnRoIiwgeyJfc2QiOiBbIjNDdXNPSmw0SmcwMmhfM2VSbW5wclRPOFlxd1VYQ3pzUS0yQTh3Ymphb2MiLCAiNzJmR1FnMlNDeXA2UDBYNWZOMTQtR21nNEpjOC1ROGtCVHY1Y2tkTFhmNCIsICJOSzZQNXJhZXhra1B0MGNhN3hXS1VhbWU5Z1dOcWItSE43Y01qak9TSXZRIl19XQ~WyJVZHlTTUc5clE2VFRNeU1nWXM1WUlRIiwgIm5hdGlvbmFsaXRpZXMiLCBbIlNFIl1d~WyI4NzVLQnE5eXRxbmNmSGk1YXVHWGJBIiwgInN0cmVldF9hZGRyZXNzIiwgIkZPUlRVTkFHQVRBTiAxNSJd~WyJJT3c2UUVUVzRVWnBjNUEtRWtrX0dBIiwgImxvY2FsaXR5IiwgIlNFIl0~WyJGdGdxVFMwMzRWaHNYRlRrckVpeEFBIiwgInJlZ2lvbiIsICJTRSJd~WyJ5cmh6dlFNMUdXblpiSVIxSkUzU2lnIiwgInBvc3RhbF9jb2RlIiwgIjY0MTMzIl0~WyJzWkEzVUt0Sy01TWZZNXhBcFJXM3hRIiwgImNvdW50cnkiLCAiU0UiXQ~WyJFa3J2RzdQRGZnYU85bnMtU05FNWlRIiwgImhvdXNlX251bWJlciIsICIxMiJd~WyJfT2dzdzlFbHhHRTdoZHlBdTM5Ri1RIiwgImFkZHJlc3MiLCB7Il9zZCI6IFsiLVlCaHNZTUViMGV0bVc2d1hzNGJCZkN4LUFjVHdkek9DWVJjMVZ3dDZzcyIsICItYlRCNzNSemI1MklpMXhsNHY1T1phbWRJWFh2UjlOM1E5R3ZzczhRMkN3IiwgIlB6YnFDRzd1NlBqSFhLU2ZDQzQ2UGxDVkdrNG1ubW8wVTRMTHRmVWdvUU0iLCAiZjlwZV9hbEM0WmxodmlSbW1CRjdLc25wbEE1czI5VHBLWXBUV0pWc0haTSIsICJnaTlpTXpQaDYxVVJ2LXV3R280N3dVTGVIMEtacFBYdFVJZGxJbnJZeDRRIiwgIm1nbHFQaFc4b05lTkV2WGF0c1BseTR4R21PZld2OUY2MDU0VndRVHpOWVUiXX1d~WyIzLWlVcTVEWnRMREwxaG5aVnNrSEp3IiwgInBlcnNvbmFsX2FkbWluaXN0cmF0aXZlX251bWJlciIsICI5MDEwMTY3NDY0Il0~WyJsUFA0clBhc0ZPem1VMXVEZ1BBSGV3IiwgInBvcnRyYWl0IiwgIl85al80QUFRU2taSlJnQUJBUUlBSlFBbEFBRF80UUJpUlhocFpnQUFUVTBBS2dBQUFBZ0FCUUVTQUFNQUFBQUJBQUVBQUFFYUFBVUFBQUFCQUFBQVNnRWJBQVVBQUFBQkFBQUFVZ0VvQUFNQUFBQUJBQU1BQUFJVEFBTUFBQUFCQUFFQUFBQUFBQUFBQUFBbEFBQUFBUUFBQUNVQUFBQUJfOXNBUXdBREFnSUNBZ0lEQWdJQ0F3TURBd1FHQkFRRUJBUUlCZ1lGQmdrSUNnb0pDQWtKQ2d3UERBb0xEZ3NKQ1EwUkRRNFBFQkFSRUFvTUVoTVNFQk1QRUJBUV84QUFDd2dCc1FGb0FRRVJBUF9FQUI0QUFRQUJCQU1CQVFBQUFBQUFBQUFBQUFBQkFnY0lDUU1FQmdvRl84UUFTUkFBQVFNREF3TUNCUUFIQkFZR0N3QUFBUUFDQXdRRkVRWUhFZ2doTVFsQkV5SlJZWEVVSXpKQ1VvR1JGV0p5c2hnek9IV0N3aFpEb2JIUjhDUWxKalExUkVhRnM4UFVfOW9BQ0FFQkFBQV9BTnFMX1pTUENsRVJFUkVSRVJFUkVSRVJFUkVSRVJFUkVSRVJFUlV2OXZ5cEhoU2lJaUlpSWlJaTRROTNfazVJX1Bzdnk3X3FpdzZWdGs5NjFMZmFHMVctbWJ6bXFxMnBaVHdzYmo5b3ZlUUdqc2U1SUgzVmlyMTZoblJwWXE2UzMxdV90aGZORTdEalNSVk5YRm43U1F4UGE0ZmdycUQxSi1pUl93Q3h2eGJqX3dEYTdnUF9BTkNfRTFmNnBQUmhwYWlmVVV1NTFWcUNwYTNJbzdSWnF0OHpfd0FPbGpaRVAtSjRWdXo2MEhTMDRmcTlHYm5menRORF93RDJMeUo5YkRhNGFpbnAyN002cGRZVzUtQlhmcDhINlVfczBfTlQ0NE03a2pfV3U4TDhtcTlidlRVZFpQSFE5UGQxbnBoSzRRU1M2ampqZTZNZUhPWUtkM0IzMWJ5Y0I3T1hZcWZXMzBULWtVWXBOaTcyYWQ3bWlxZkxlSW12WUQ1TVliRTRQeDkzTlZ3Wl9XUTZXWXFsc0FzZTRVekJISEktYUMxVXZ3Mmx6V3VNZVgxSWNYTXlXTy1YQmN4eGE0dHc1M3Z0TGVxUDBZYW5qZ2JOdW5QWTZtVnZlQzZXZXNpTEhmd3VlMk4wWV9JZXNrOUY2NzBsdUxZYWJWV2h0VFctLVdpckdZYXVobmJMRTdfaWJudjlzaGVqUkVSRVJFUkVSRVJVdjl2eXBIaFNpSWlJaUlpaklYNVY2dnR1MDdiS3E5WHk1MDF2b0tDQjlUVTFOUksyS0dHSmd5OTczdU9HdGFPNUpJQThFbGE1dDV2V1kwcHA3VUU5ajJXMEJfMG9wYWFWMFp1dHhxWDAwVlJ4ZGpNVVFITXRKX1pMaTBrZVdyRURmRDFOLXB6ZUNwTU5zMU0zUlZvYzNqLWdXRjVqNWZkOHpzeUgtUmFzYk5TN243azZ5cFA3UDFidUhxUzkwalpQaXRndU4xbnFZd18tSU5rY1JudjVYbUNTZXhQaEV5U01FbkFRRWp3U25KeDh1UDhBVlR5ZF9FZjZweWQzLVk5X1BmeWdlOFl3ODl2SGZ3bzVPLXBWeWRwdW9QZW5aS3NNLTFtNGQzc1RYU2ZFa3BvSlM2Q1FfVjBUZzVoUDVDMkpkTm5yRVJ6UHA5TWRTMW1GTzQ4WXhxSzBVNWMwRXV4eW1wMjUtVUR1WE16X0FJWExaYnBUV0dudGIyQ2kxVHBDX1VsNXM5eWlaTlNWMUZNMmFDVmpzOXc5cDc0OEh1RGtZX2FCQ19meUZLSWlJaUlpSWlLbF90LVZJOEtVUkVSRVJSa2ZWZEs0M0tpdGxITGNMaGNLZWtwSUdmRmxtbmxER01aX0U1eElEUjl6MldJUFVKNm9QVHhzMDJXMWFTdlROZjZnQWNCVFdXcGEtbGhlT3dFdFNBWXdNLXpDNF9YaXRQdThQVXp2UHZkcUM2M3ZXdXZyMUxUM09iNHB0Y2RkSXlpaVp6NVJ4dGhEdUJEZnFSbFdseWZHU2hKUGtuc2lJaUlpSWlaUDFWVFh1Sjd1SjhuejlmS3Z2MHpkWGU3X0FFd2FnaHVHakwxVVZGaWttRHE2eFZVaGRTVkxCLTNodVBrZWNmdE53Ujc1VzdqcHY2djluT3AyeENzMEJxTmtkN2hqWTY0V0tzSWlyYUozRnZMTENmMXJPUkFENC1RUElBbHJzaHQ5R09MaG5QbndxaDRVb2lJaUlpSWlwZjdmbFNQQ2xFUkVSRkI4SzEyX21fbTMzVHR0X1hiZzdoWGNVOU5TakZOU3h2QnFLNllfc3hSTThrbDJBZlpvQkotVUZhWk9zVDFCOWZkVmRKVDZVbzdRTks2VXBablN1b1lLb3ZmV09QWUdkX3l0T0dfdWdZeXNSblB6NEtqazdPZVJ6OWNxRVJFUkVSRVJFVTgzWUk1SEJ4a1o4cjl2U3VyTlM2SHZ0THFYU043cmJOZGFKd2tncTZTWjBVMFo5OEVISkI3NUh1c19PbXoxZHR5TkpWOUhwN3FBcFJxcXhrdGpkZEtTSmtkeHAyLTczQm9ESmgzT1I4cmhqc1Z0ZTJ6M1IwSnZGb3VoMTd0dHFpbHZsa3IyY29xbW5lNXBZV241bVBZNEIwY2pULTB4N1E0SHNRQXZZWmI5VlVpSWlJaUlpcGY3ZmxTUENsRVJFUkY0dmRYZExSLXpXZ0x6dVZyMjhDMzJheVV4bm5lN0hLUjJRR1JSdFA3VWtqaTFqV2VTWEFESGNyUWoxazlXTjE2cnR4bzlXU1dSMW1zMXVnTkxiYUF6dWt3M2s0bVJfZmp6Y0NNaG9BR0JoWTg1SThFb2lJaUlpSWlJaUlpWk9NWlU4bllJNUhCT1NNcS1uUzcxYWJsZEstc2Y3YjBWWEdvc3RkTEhfYTFrbWNUVFZqQjVJSGxrZ0dlTHgzR1JuSzN2OVAzVUp0NTFKYmUwZTRlM1YxRThNMkk2MmllUUtpM1ZBQUxvWm1lV3V3Y2ctSE40dWJrT0JOMGc1cmhrRlZJaUlpSWlLbF90LVZJOEtVUkVSRjFwS2hrZVh1a3dHLVNRY0Fmbk9NRHlTZlphU19VOTZ2bmIyN2lPMm0wVGRYUzZLMGhVUFktU25lVERYMXd5MTh2YnM1ckFYTmFUN2x4LWl3VTVIT2NuUDFSRVJFUkVSRVJFUkVSRTVINmxYbjZZZXAzWF9TM3VGSHJqUnRRYWlsbmFJYnBhcFhFUTNDbnlTUTdIaHdKUEYza1pPRnZxNmMtb0hSUFV0dGJRYm9hRm5kSFQxRGpUVjFIS1FacmZXTURUSkJKanNEaHpIQS1DMTdYZUNBTHJvaUlpSWlLbF90LVZJOEtVUkVSRnJkOVVQcmVPZ2JUVjlPMjFsMWxoMUxkSWZfYUM0VTctSm9xUnpEbW5hN3lKWGpISS16UGw4dUhIVDJaSFpKOEUtVlNpSWlJaUlpSWlJaUlpSXA1T0hoeF9xc3BlZ3JxNnV2U3p1ZXlTNHpTejZLMUctT212bEtDZU1CemdWVFJfRXdlUjdzSl9oQzM2VzI1MGQyb0thNTIycVpVMHRaRXllR1ZqZ1EtTjRCYTRZN0VFRUgtYTd5SWlJaUlxWC0zNVVqd3BSRVJGMUtxc2hvS2VXc3E1MnhRUXNkSkpJOTJBMWdCSkp6MndBQ1NUOUY4MWZVSHJhUGNYZkxYV3VLT3NkV1UxM3ZsWlVVMF9mOVpBSlNJM2o2RGcxdVBzcmJJaUlpSWlJaUlpSWlJaUlpcUQzZTN0Mlc1bjBqZXBpdTNKMjNyOWtOVlZqNTd0b1dKa2xybmxlQzZXMlBQRnNaOV8xVHNOX3dBTG94NEMyR29pSWlJaXBmN2ZsU1BDbEVSRVdJWHFpN25YdmJEcEp2NzlQMWJxV3IxVlhVMm5ETzEtSE1obmE5MHdiNzVkRkRJMzdDUng4Z0xRcTRueDdmUlFpSWlJaUlpSWlJaUlpSWlJc2t1Z3pxRWg2ZE9vZXhhc3VsVndzTjJZYlBlWFBHR3gwMHp3QklUN0NOd1pKZ2VjRmZRWlNWY1ZaQkhWVTBva2dtamJKRzlqdVRTMGpJY0Q3NUJCWGFSRVJFUlV2OXZ5cEhoU2lJaUxVOTYwdTdzTlpkTkU3SlVUbmwxdkVsLXVBejh2S1JwamhianhrTjU5X3dDLXRYQ0lpSWlJaUlpSWlJaUlpSWlLUTV3OE9JOV9LMy0tbkp2MVFiNWRORmhhNzRrZDYwYkhGcHk2eHlTYzN5U1F3c01kUUNlX0dWaGFlX2h3a2I0YXNxa1JFUkVWTF9iOHFSNFVvaUlpMGJlci1UX3BjVERQalQ5dl93QXJsaEFpSWlJaUlpSWlJaUlpSWlJaUxaRjZLMnNxbTNidWE2MEJMV3NiUlhxd3hYRmtMblktSlBUemhnTFI5Zmh6U0VfWUxjS0NENEtsRVJFUlV2OEFiOHFSNFVvaUlpMGJlc0FUX3BjemR2OEE2ZnRfLVZ5d2dSRVJFUkVSRVJFUkVSRVJFUkZrRjBLYm50Mm02cDlCYW1xNjhVbEJVWEZ0cnJwU3d2QWdxbW1FamlPLWZtQXo3ZVY5RURNZUd1T0NCajhmVUVMbVJFUkVWTF9iOHFSNFVvaUlpMGJlc0FSX3Bjei1mX2dGdl95dVdFQ0lpSWlJaUlpSWlJaUlpSWlJdVNubm5wNW81NmVaOFVzVGctTjdIRnJtdUI3RUVlQ1BxdnBkNmZOYlUyNHV5bWlOYjB0UTZWbDRzVkpVRjdoaHhjWTJoLWU1LWJrRGtleEpIc3Jqb2lJaUtsX3QtVkk4S1VSRVJhTnZWXzhBOXJtYl9jRkRfbGNzSUVSRVJFUkVSRVJFUkVSRVJFUlMzeXQ0M3BCYXRPb2VrZGxqZkU5aDB4cVM0V3NjejJjMTRqcXdRUDNSX3dDbEh0OVFTczRVUkVSRlNfMl9La2VGS0lpSXRHX3JBZjdYRTMtNExmOEE1WExCOUVSRVJFUkVSRVJFUkVSRVJFVXQ4cmE1NkkycGJ1LTNicWFQa3F5NjEwMDFydWNGTzVvNHgxRXJaNDVYTkk4bDdJWWdRZkh3eGo5NWJTMFJFUkZTX3dCdnlwSGhTaUlpTFJ0NndQOEF0Y3pfQU80TGZfbGNzSUVSRVJFUkVSRVJFUkVSRVJFUkZucjZQLTVkWm9fcVJyOUJUT2xGSHJXelRRRnBJREcxRk1maXh5SEl5U0dpWmd3Y0Q0aEpCd3QxN0hsMlNleGI3TGtSRVJGU193QnZ5cEhoU2lJaUxScjZ2emc3cTdxR19TdzJfd0R5dVdFS0lpSWlJaUlpSWlJaUlpSWlJaXZuMFdzckg5VjIxa2RESzVsUWRTVWdpZjI3SGxqN191Z2pCOGVjcjZNWXVPTWc1SlhJaUlpS2xfdC1WSThLVVJFUmFQUFdBczkwby1ySjF5cktiNGRKY05QME1sSV9PZml0WmxqajlzT0R2NkxCcEVSRVJFUkVSRVJFUkVSRVJFVXM3bnVzN1BSXzBkYjlSZFUxWGZLMFAtUHByVE5YY0taalFDSHVmSkRUa2tFZnVpWTR3UWVXTTVHVnUyakFHY0JjcUlpSXFYLTM1VWp3cFJFUkZxUDlhUGI3VXJ0WWFJM1dkQy1Td3V0eHNJbUFiaU9wRWtzN1d1N19BTHpYT0k5dmtKLWdXc2gzbFFpSWlJaUlpSWlJaUlpSWlJaWVQQzJpZWlEU1JTM19BSFpySkdfcklxTzBSTno0NHZkVWtfei1RZjFXMlBBSGdCU2lJaUtsX3QtVkk4S1VSRVJXTDZ6Tm9YNzU5TmV1ZHZLV056N2pWVzhWdHREU0duOUxwbnRuaFp5UFlCNzQtRG5lUXlSeEhjQmZPcFUwMHRMTkxTMUVMNHBZbkZqMlBiaHpYTkpCQkhzUmdnajZoZFpFUkVSRVJFUkVSRVJFUkVSRkxPNTdyY182UlBUN3JMYW5iYlVlNWV0S0NlM1AxOC1oTnRvNVFHeXNwSUd5dWJNOGNzdC1JWnNOYVc4dUxXdThQdzNZUWlJaUlxWC0zNVVqd3BSRVJGUTRFOXdGODVfV250OE5zT3FMY1BUQTVfbzV2VTFmQVg0NWNLaHhteUFBMEVmT1FyRklpSWlJaUlpSWlJaUlpSWlJc2plZ3pZU2s2Zy1vN1Qta3IxUWlwc051RDd2ZDJ1aTV4eVUwT0NJM2prM0RaSDhXRV8zc0w2RWFLaXA3ZlR4MGxKVHh3d1FNRWNVY2JBMWtiQU1CclFPd0FIc0Yya1JFUkZTXzJfS2tlRktJaUlpMWstc1hzVG9wbWhiWDFCMFZLLW4xSTI1MHRrcW54bjlYVXd2am1jQzhlN2dXQUFqdmpzdFJxSWlJaUlpSWlJaUlpSWlJaWx2bGJqZlJ5Mk9icGJhNl93Qy1ON3RMNHJ0cTJzZGJyYkxLekJGc2d3UzZNOGVZLUpPWEIyU1E0UVJrZUR5Mk5JaUlpSXFYLTM1VWp3cFJFUkVXQzNyRTkta2VGekI0MVpic19ZZkJxUDhBeEMwZ0lpSWlJaUlpSWlJaUlpSWlMbmhpa25rWkZER1h5dmNHTWExdVNTVGpHUGNra0FMNllkaTlHVUczdXp1aXRGVzZtWlR3V2V5VWRLeU1Cd19aaEFjZm0tWXVKSkpKLXF1QWlJaUlpcGY3ZmxTUENsRVJFUllsLXFIcG0zNmk2SzlkMUZZek10bGt0MXlwSGdmNnVWdFhER1Nmcnlqa2UzN2NzLXkwRm9pSWlJaUlpSWlJaUlpSWlLNE93OWhmcWJlblExaWloZlVQckxfUVJ0aWJMOEl1QW5hWGZOeGNCZ0FsZlM1UjA1cDRZNG1kMk1ZRy1BTTRYYlJFUkVSUWZDbEVSRVJGYjNmZmJPaTNrMmUxanRmWFNOalpxVzAxRkN5WWpQd1pYc0ppbHhfY2VJM19BSDRyNXI3N1k2N1RGX3VlbmJ4VGlHdXRWVE5SVlVKSVB3NTRuRmoyX2ZpOXA3LTRIM1g1S0lpSWlJaUlpSWlJaUlpSXI5ZERGcm11dlZ0dGZUUjB6NWdiX0E5X0JwZHhZME9jU2VQc01kMTlGWVlBY2tES3JSRVJFUlFmQ2xFUkVSRlFXTTg0N0xTOTZxM1NUSHROcm9iNmFQYTV1bnRaMXoyM0NsYTA0b3JnV2h4ZGtfdVNuazRmM2c3N0xYMmlJaUlpSWlJaUlpSWlJaUxNbjBwZE5WZC02d3RQMThkSExOVFdXMzE5ZFV6TVlTMkFmQWMxaGNUN0Y3dzM4a0xlOTkxS0lpSWlLRDRVb2lJaUlpdGwxQjdQV0Rmelo3VTIxV29JbUdHLTBUNDZlYkdUVFZiZm5obWJfZVk5ckhBZTRhUWZLLWRuZFhhcldHemV1THBvRFhWb2ZRM1MxekdON1Qtekt6bGhzakNlem1PSGNILXE4VWlJaUlpSWlJaUlpSWlJaTI4LWpWc1hkOU02UTFQdmpxRzIwMExOVkNHMjJPUWdfcERxV0NXUTFEdTQtV044b2lEY0g1akNTZXdiblplaUlpSWlLRDRVb2lJaUlpcExXLWNEUDFXTTNXdDBkNlY2cDl2S3FDS2pwNkxXdHVoZk5aTHFXY1h0ZTF1ZmdTRWQzUnZfQUdUOUNRUjRXaGJYT2hkVjdkYWxydEk2MXNGVmFMdmJablFWRk5VeDhDSHRQelk3Zk1NanNRZkJDODJpSWlJaUlpSWlJaUlpbmlWZURwcDZkTlk5UzI2RkR0NW82M09rZy1JeWU3VjNMZ3lob3c0ZkVrYzhod2FjWkRmbGNTN0FISUVyNkl0RGFNc1czMmtiTm9mVFZGSFMydXlVa1ZGU1FzWUdoa2JHdGFPdzdBNEJQODE2VkVSRVJFVkxsVWlJaUlpSW9JR01ZQzFfLXJ4c1c3WEd4ZEh1dHAtMFJ5WExROWFKYTk4VVEtSy1nbEFqYzRrREpFYnhHZV9ZTkx6N0JhV2Zvb1JFUkVSRVJFUkVSRVdYUFFQMFRXenEtdU9yVHFYVWx5c1ZvMDVUVV9Hcm9JV1BMcW1WNXd3aDR3UndZOG5IY2R2cXR2dlREMG5iWWRLMms1OU42QWhxNm1wdUVySjdsZEt4NGRVMXIydElia05IRnJBQ1NHdHdBU1NyNDhXLWVJX29wUkVSRVJGU193QmxJOEtVUkVSRVJGYkxxVnNkWnFUcDEzUzBfYktYNDlkYzlGM3VrcDRnM0pmTkpRek5qQUgxNUZ1RjgxSlpqQUlHUm4tZVBiSHN1SkVSRVJFUkVSRVJFVTRKT01aSi1pMzNlbUxzNl9hVHBTMF9VVjFHNkM3NnhtbDFKV3RrYmd0Yk54YkMwWjc0LUJIQ2VKOE9jXzZsWmNZQ2xFUkVSRVJVdjl2eXBIaFNpSWlJaUl1R2VKc2tUNDNOQkR1eEJHYzUtcS1mVDFDTmpyUHNKMU1hZzB0cGtCdG11ckk3MVJRal93Q1haUG5sQ1BiRFhCMlBvQ1Bvc2FFUkVSRVJFUkVSRVJYUzZaOXRaZDJOLU5ENkNaUV9wVU4wdk5NMnFqZEVYc05PMTRmS0hZOE5NYlhEN1pYMGwwZEhTMFZMRlFVY2JJcWVDTVJ4eHNIWnJRTUFENkRIc3UyaUlpSWlJaXBkNVZTSWlJaUlpS2w0SkhaYWt2V2syM3FLZldPaHQxYVdpZjhBQXVGRlBaNnVZTU9CTkc3bXhwSTktRDhEUFk0SzFqdVVJaUlpSWlJaUlpS1dyYVQ2Tk93MVBVeWFvNmg3M1NCNzZPZDJuYktIanUyUXNhLXFsd2ZCNHVpWm43eWU0VzEwQVk4S1VSRVJFUkVVSHdwUkVSRVJFUlFmQ3RwdjFzdnBEZjdhMjliV2F4cEEtaHUxTTVrRTdXZ3kwVlExdjZxZVAtRjdINElKN0hCYTdJSkItY2pjSFEyb3R0TlozblFPcnJmLWgzaXhWajZLcWlJOFBhVDNiX0UwOWlDZlloZWFSRVJFUkVSRVJFWDZ0aHNWeDFMZXFEVHRsbzMxTnd1VlZIUjBzREc0ZExLOXdhMXY1TG5BZlFkeXZvODZZOW5MZnNMc1pvX2E2a0RETlpyY3h0Yk0xaGI4YXNrX1cxRW1EN09rZTRfWVlIN3F1dDRDbEVSRVJFUkZCOEtVUkVSRVJFUlVQYU1mOXEwSy1xalJVVkYxbGFwZFF3eHh1cXFPaG5xT0FBNVNmQkE1T3g1UEZyZTVXSWlJaUlpSWlJaUlpenc5STdZLVRjUGZ1cDNKdU5HeVN5NkVwdmpmck83WDFzN1hNaGJfQUNEWHZfNFI5VnUwWTFvYUFHZ1p5ZkgxOHJrUkVSRVJFUkZCOEtVUkVSRVJFUlV2X1pYel93RHFhVlV0UjFuYTliSWY5US1raVpuLUVVMGF4WlJFUkVSRVJFUkZMZkszVmVqbHR4WDZWNmNydnJxNHhpTnV0YjI2YWlCSGQ5TFNqNEh4UHRtVVRBRC00RDdyUHdlQXBSRVJFUkVSRkI4S1VSRVJFUkVSVXVXZ0QxTktTYWs2enRlUG5hUi1rUHBaV1pfaE5Pd0RfS3NXRVJFUkVSRVJFUlN6T2UyUDVyNkNQVGUxbllkWWRIdWdXV0dsX1JXMldrZmFhcUk0ei1sUXZfV1B3UFp6bkZfZnVlV2ZkWlBEd3BSRVJFUkVSRkI4S1VSRVJFUkVSUWZDMFYtcmRhWmJiMWdYS3BkRzRzdUZrb0tocmkzQUp3NXVBZmZIRVpfS3dzUkVSRVJFUkVSRkxXOGpoYndQU0YwWmVOTDlLRXQtdVRXdGkxWHFTc3VsRTBGMlJBMWtWT0M0RnZZbVNua1BiSUk0bjNLem1IaFNpSWlJaUlpS0Q0VW9pSWlJaUlpaDNMMkFLMXVlc0QwNzNEV09oclB2NXBxa004LWttbWl2TEdOeTRVRWpzeHo5ampESGtCeEF6eGVENEJXbnhfWTlsQ0lpSWlJaUlpQUUtQXZaYlZiYzN6ZG5jS3c3YzZaaEw3anFDc2pwSVNBMXdZMXgtYVE1TFFBMEJ6amx6ZXpjZTYta3JiTFFkbjJ2MjkwOXQzcC1QaGJ0TjIybnRsT0R5eVk0V0J1WEYzY2s0Sno5eXZXSWlJaUlpSWlLazltOTFVaUlpSWlJaUl2eU5RV1cxYWxzOWRwNi1VRVZiYnJuQkxTVmRQTXpNY3NVakMyUmpnZkdXa2otYS1lcnJKNmFybjB5NzJYZlJQNkxVTzBfVlBOYnAtcmthVC1rVWIzRXRieVBaejQtN0hEM3dEN3F3aEdDb1JFUkVSRVJGeXNhU2NOSGM0eDlUbndBQjdyYnA2VUhSM1hhTnRNZlV4ci1ra3A3cmU0SHc2ZnQwOU81ajRLUjVIS3JmeWJ5RHBBQ0dZd09CNVpjSGdOMlZzajRqdzNJOThMa1JFUkVSRVJFVU84S1VSRVJFUkVSRnh1akJka0J2OVB2bFdLNnQtbHJTUFZUdHRWYVB2Y2pLSzlVYlh6V1c2dGo1dm9xb3RJSEpvN3VqUHk4bTV5UUFXOXdGb0szYTJoMXpzdHJ1NWJkYmdXVjFCZUxiSVdPYVBtam1aakxaWVhqdEpHNGZNSER0Zzl3MGd0YjRWRVJFUkVSRlcxdkk0d001OXg1UDB3RnNZOU9EMC1uYm1WVVc5dTlsbHFZTk1VRTdIV1cwVlZNLU0zYVFBT0ZRNXIydDVVdzVBTkl5SHU1QUVCcHp1R3BxU25wWVdRVThUSTQ0d0dzWXhvRFdnWXdBQjQ4RC1pNWx3enpmQmpNandRMGRqMzhENnJtQkJHUXBSRVJFUkVSVXVWU0lpSWlJaUlpakFWQmpqem5IOGxqSDF2ZEcybmVxX2I2YUdnTkRidGNXbU56N0ZkWm11NE9lMDVkVFRGbVhmRGZod3lBNHNjNE80dTRscnREMnRkQmFtMjQxUGM5RjY3c2RWWnIzWjVUVDFWRlVSa09hNXBHU0hESWNPLVE0Wlk1dUhOSkJibnpDSWlJaUl1d3lCMHJteHhNTG55RU5ZR2pKSnpqeDV5ZllCYktQVC05TlM1YWxyNmJlUHFKMDVQUVdhbmw1V2ZUbGJDWTVxOXpYWU0xUkc3RG80OGdnTWUwT2VCeXh4TENkdU5QU1UxTERIVFVzTEk0b21OWkhHMW9EV05IZ0FEc0FPMkF1ZEYxYmhFSjZPYVB4eWF1ZVA4QVpINFZhSWlJaUlpS0Q0UWVGS0lpSWlJaUlpS2t4dE9jak9leFdKWFhMME82YjZxdEtTM215Q2x0bXY3UERfNnF1Yl9sanEyZ2wzNkxVT2FDU3dqczF4QkxDUTc1aHlhZEdXdGRDNm8yNjFQY2RGYTIwOVZXZS1XeVoxUFZVZFV3dGtpY0RuSVBod3gzRGg4am1fTTBrRUZlYlJFUkVYdHRzZG85dzk0OVUwLWk5c2RJVnRfdkZWLXhEVHNEV043SExueVBJWkczQU9YT2NBQ094VzRYbzQ5TWJRZXhYNkpybmRlYTM2eTFxMTBkUkNHd2swRnFrRFc0YkUxM2VWNGVYZnJITkhnWWF6QjVaMU5pWXp3M193QV8tUUZXaUxwM09vYlNVTTFTNXBJWTNPRlhSMUxLcW1qcUlzOFhESzdLSWlJaUlpS0Q0UWVGS0lpSWlJaUlpSXFUR3duSkhkWXpkWEhSRHRqMVcyZVNwdTBMTEpyT2twekJiTDdUTlBKdVNITlpPd2RwbzhnOXZMUTV4RGdWcEgzNzZjTjB1bTNXTXVqTnpMQ3lubEl6UzE5TVhTMFZjenlId3lrQWtaeU1FTmNNZk0wWlZxa1JGLTlwYlJPck5jWEtLeWFNMHZjNzVjSmp4anBiZFNTVkVyX1A3ckFUN2ZSYkQtbkQwZXRZWHVzb3RUOVJWNXByTmFYUnh5aXdXMlYwbGJMemJrc25sRGVFSEFrQWhwa0xpSEFPYUFITzJmN1ZiTjdaN0o2ZFpwYmJEUnR2c0ZDM0hKdE5DQkpNY2s4cEhrYzVDTW51NG55VjdveHNKeVI3WV9rcWtSRi1kZTQzU1cyWmpTY3VieHdwczBUb2JkRkVSLXg4by00WDZDSWlJaUlpS0Q0UWVGS0lpSWlJaUlpSWk0M01aX0Mwa0hPQ1BmT2YtOWVLM1IybzBEdk5wT28wWHVMcDZrdk5wcVNDSXBobmc4SExaR094bHJoN1lXcWZmMzBlOTE3RGZhbTQ3QVZ0djFSWkpCemd0MWRXeDB0YkNQNGVjaEVjbjU1TktzSGZmVGE2MXRQVWJhMnYySXVFckM3aUdVTnlvYTJUSDE0UVRPZF8yTDBHanZTeTZ5ZFV5MElyOXVyZHAybHJUM3FyeGVLZG42TzM2dmlpZEpObi02SXk3N0xMX2FuMFg5dXJMVVJYSGVEY201YWtEVEdYVUZwcHYwR0RfVnU1dGRJNXo1SGptUVd1YndPQUFSM0t6XzBGdGJ0eHRmUU9vTkE2THM5aGhrSDYwMGRHeUYwcDc5M3VBQmQ1UG42cjFvQTg0R1ZKYTBqQmFNZU1ZVW9pSXV0VkJyZ0l5YzhqbGM3STJ4dERHakFIWUtwRVJFUkVSRlNfMlZTSWlJaUlpSWlJb3lQcXFlUUhrbnpqdjJ5Zm9zTWVyUDFMZHItbTY3emFGc2xvcU5YYTBwSldzcmJmRTkxTkJRc2Mza0hTVEZqZzUyTVlZMXBKemtsb3dIWUVXMzFPLXFQV2U1dHVscjdwU0d5M0s0dzA3ckhRVWtiSXkxNzJzY3lOMGhCNUhtT1B4SGdaeGx3R1ZtbG9uMURhdDljYVBVMmk3N2JxUDlOam9LZXN2VkFhS09zbmNTR21MazRTX054eU9UR2o1bXRQRnhJV1ZPM21fV2dkdzZpTzJXMnVkUzE4ak10cEtnY1hIN05QaDVfQzg5djMxUzZGNmNHMDlmdWhicnJRMlNzSWlodWtGTy1lSjB6c2tSOFdOYzRIRGZjWVhrOXBfVUk2V3Q0THJGWk5QN2lRVUZ6bWtiQkRUM1dOMUc2YVJfTGkyUDRvYVhFNDc0QkE3SzdPdGQ3ZHI5dDd0YjdOcjNXdHQwOVUzaVJzRkE2NHlOaGpxWlhmc3RZNGtOSi0yY19kZTZqbWJJeHBqZnl5QWZsUFlnLUQ3OXZ3VnpvaUlxU1RuSHNmUDJYV2o1U3pseEdHdEhGcTdROEtVUkVSRVJFVUh3cFJFUkVSRVJFUlVPZTBaeWNjVjBwcnJSUjFMYUg5SmlkVnZZWkdVNGxBa2N3RUF1QXprZ1o4NFVWZEJIWGlOOGtrelBoUDVzREpDMGgzMU9QUC1Id3JOZFFfUjFzbjFMMFViZHc5TXhpNjB3NVU5MW9jUVZqUGx4eGRJMFprWl9kY01mWllmVjNRcHBEWXJVOXRyNld6c0ZOREhJeUtzZUJVTXJIY2c2Tjd6SUNXU015Umx1TWc5dVEtWmVtbTBZWkh1Y3lqZkxHNDVBOThFNUJ6LWU2NzFtMFhkTFRYTnVWcmxmVHpNbGE2SjBiaGxwNWVjX3dfZnlzZ3lORTlRdWdhX2FmZUswMFZZS3lFMDdoSTlnLUs4dExSTEE0amt5VWNzZzhRUjlGcDU2MXVpclZYU2RxMXRSSEktN2FMdk5SUF9ZOXdhd2gwUUQzRVU4NHljUEVmRDVzNGQzY09QN0t5ejZJdF93RGJycXkybmw2UC1wZUtudUZ3Z3BoRFlxMlp6M1QxRUxJd1FSTVE0TXFJOEF0SmR5a0p3UjJPZTNzcjFLYmo5QnU4c3ZTdDFUWDkxMjBZOTRuMDlxaDhqNVAwYWxsTGhGSzRubEo4QnhhV21Nak1MbXVQSnpCazdON2JkS082MGNGeHR0ZEZXVTFVeHNzRTFPOFBqa2pjQVd1YTRFaHdJNzVIYkFQZ3I5RkVSY2J3UzF3QjhxcGpBMXZIQVZTSWlJaUlpSW9QaFNpSWlJaUlpSW96M3dzYWV0VHJEMDUwbmJkUlhpV2piZHRWWHh6NmF4V3draU44Z0FMcHAzanMyS01GcElCRG5PSWEzQTVPYnBpZzZ2dDluYjZXenFFdS11SzY1YWx0VlMxOGJaWkN5bmZTX0U1dm96R3dnTmdmZ3NMQjRCR0R5YUN0N2ZUX0FOUU9oLW9yUUZ0MTVvZTROa1pVc19YVXJpQk5TeTRBZkU4Y2p4Y3duRHM1OWlPUWMxeHV1Q1BZZmZ1dW5jYlhRWFdqa3Q5eHBZcWlDVnZGOGNneUMzX3hWb3RTN1AwMWtZNjRXZHNrOUZHMGwwSGx6SXdjX3dBeGozOHExTjVyYWVra2xiQTBrUkRQWUh0OXNIM1hncV9WTnlwS193Q0pRel9BbHhoazRIenhqN0Izbi1hdkpvdlhtaTktTkxWV3ltN1Z0cHJxMjVVajRwV3pzNVIxTFFjQTh2M1pNOXdSM1dvZnJENmE5UTlITzk5TFJXTzUxWDltVFBiZWROWFlFZkZqRWNtV2prRGt5UnU0RGtRMzkwLTVWMWQtTi1mOU5ucGJ0ZDd2TmhmSnVqdHJYUnczRjlCVGlULTByZE13ODZvTllESkd3R1BMd1FHQjJIWi1ZTmJ6LW03MXJYall2WE5Gc3Z1SGVKMzZBdnRTWUthT2Roa05wclpYNEQyWklkSEU5N3N2R0NBVHl3MzUzSGRkSE15b2EyU0tVUFlRQ0hOZG5JUGYyN2R4M0JDN0NJb1BoU2lJaUlpSWlJb1BoU2lJaUlpSWlLRDRYbmRSYWptb0hDZ3RkTS1vclpld0FIeXhELUluM19DeHozQjZPTkQ3dDZocTlUYm9XQVgtNVY4VVVSZlhTT1B3bXhod0FpSElCZ0JlNXpnTWNpUVQ0V0VHNGZwT1hHMDZpclJwM1hFOU5iNVdFMFVkVlNmRmZFNFBiMkx3NXZOb1lTT3plUUk1SHNEbjlQb1dxZGViSjZtMVB0RnFHT3Z0MTYwOVV0cTNSbDdIMGRTeVlfRGpraU9BOHRjR0QtNzJ3NE5jSEFiUWR1OXlxTFd0QzBWTEJTM0FPTEhSY2dROHQ4bHVmWmU0Qjl2Ql9xcVh4aDRJTFdrSHlDUEtzSHZmdDFQYjRLalZWamFYUVBEblZFQXp5RHljaDM5ZXl3NDNDMXphTkQyMm92bDl1VVZGQkUweUV6WWE5ekFQUEVuT1Q0NGprNG45a0ZZbWExNjJiMmF1S3IyeXBxeTFWMUxVUlZFVmJPNXJYdExDN2t4MGVYWkdDMGdoelNNSE9WZmZmemVLMTlkWFJJZGV6VWNWRnVEdE5jSUo3MVNqTFd6VXMzNmt5dzR6ekR2a1BFa0VPYTRlTWNybGVuSjBNNjEwVnAyWGRyY2R6N1hVNnJwR01wYk84WW1ocGVmSnNremZMWkhIRGhINURjQndEaVEzTGJYdlJic1Z1TGFhaW4xSG9XMFZsd3FIdGxrcm4weldUT2V3QU1jWHNfV0VZQUJITDIteTQ5a2J4cXZhNjlRN0Y2M05aWDItam8ydTA1ZnB4STdsVHhsc1pwS2lWd3daV2dndGR5eTVwSUk1Tkpka0tIZUQ3SDdxcEVSRVJFUkVSRVJVdVVqd3BSRVJFUkVSVVBhWHNMUTRqN2dyZ3A3ZlMwdnpSUkFQUGZrN3VmNnJzY1YwTHBhWUxwQ1k1Z0FSMlk0RHVGaVAxUGJPVVdsdGFhZjNxczlETTZSd2ZZN3QtaDA3bnZrRTVEWUpaUUhENUdQQmJuaTdIUFB5c2E0aG8tNVZYNlJETFJ1NFNSNEVRaGR3SXo1N2xaSDZCMXpfQUc3UnNwYm05amF0ck12ZUhEQi14eDRLOXVEOWUyZkdDdXRYVWROWFUwdEZXUk1raG5IR1JqaDVCOGZ6ejNfSzBrZXFCdDFVN2NiaFdtMFB0YzhsTkkyZWVpdWppWE5rZ0phVEVmYm1IWkpIanpoWVBPY19KSEozYkh2OVBDMlUtanp0SnFXNTZzMWh1TlgydUoya24yOFdPWXlPNWlvbkpaSndEQUQteU9KY1RqR1FBUFBIYnBGVHd4d3RpaGFBd000dGJqdGo2WVhLWTJrRUFZejlGNV9WdWxLSFZGdGRUVHQtSE5IODhFdzh4U2ZVTHE2TXUxd2JHX1R1b2NDNTBBd1hBWUU4WDdyeDlUano5MTZuTHMtRldpSWlJaUlpSWlLbF9zcEhoU2lJaUlpSWlLTUJTb3dmcXZ4ZFU2Y29OVldDdjA3WHRCaHVFVG8zbkdjRXR3SGZrRUFnX1pZbjI2eFhEUWwtbXNOMmdJa3A1Q3dGMmVUbXRfZXg0NGtZY0Q1OGhmdmFtcnFlZ3NUNnAxUmlXSGxJd0RMTU9IakJIMVg3M1RfMU4yZlhsU05LWGk2VTBseWMtUmxKVXRmbHM3Mk9MSHhIc0FITmVPR001THdXLVYzOS10X2FQYnlHdXRjMWJGUVZWRzAxTHBaWk9MR1JCcFB4ZVo3WWJndUpPQU9CQkktYk9JbldYZW91cXpvdnVPNDlydDhrR29kdDcwYVctVXJZSHU1UnhPTERNM0lCLUU2TjdaZzRGelExeHdUamtzQy1rM3BWMXQxVmJpeDZWMDlHNmpzMUZ4bXZkMmN6TEtTRTktTGM5blN1N2hvX0xqMkFYMEE3Y2JiNk8ycDBaYk5DYUhzc0ZzczlvZ0VGUERHM2oySGtseDd1Y1RseGQ1SkpQdXZXNEhqQ2xVa0R6Z2R2Q19NdU5taXFxbUs1Ukg0ZFZCM1pKOVJfQ2Z0OWwzWUpYeU5fV0RpOGZ0Tl9oX243cnNJaUlpSWlJaUlpcGY3ZmxTUENsRVJFUkVSRVJFVWNXNXpqdXJDNzBYLXgyX2RIUW1pWDJ1cnFyeHJlV3BwcVo4Rk9Ic3AyVXpQaVBtbVBKcGJHQWV4NzVKd3ZIOVJtbm5hQzI1cnRVYWdtZEZaNktKMVJWUG93OTdxWU5ieWNTQTNMbThRUU1kOHJTbnJqZEp6dDFhclhtMTFUWDJGckhPZFRTaDdtVGx6dWI1SHVhWE9EUzU3M0hBLVZwd1FGc2xndnRuNjRlbUhTdTVkNnQwTjAxVnBDV0doMUZSOFM3NDA4SkFaVWNQbC1VOGdfOTVnRDNBazkxY1RwZzZkclRTYUgzTzJiYUtxbTA1cTIzX0JkVHRieGpwakl6NGJ1QkhjdXlTOHZkbDNkb1BGdkVESmpwODZlOXVPbV9ROGVpTnU3V1lJSEVUVlZUSmwwdFZOeGExejNQOXllT1FCMjdsWFVER054eFlCanNNRHdxa1JVOGZ1cHdQT0FwUkVSRVJFUkVSRlNfd0J2eXBIaFNpSWlJaUlpSWlqSVhFNTdJMk9rZV9pMXVTNGs5Z0I1T1Q3TEdiWXZXT205LTlfZGVicVd1T3Rtb2RCdWRvMmdubmplMkNTVVBMNnFXSG0wWkxYTVl3dTktSVB1T1g2SFZWdWZiTEpwUTZaTHBwWnE5NWlheUdOMG95R3VjUzRnZkt6NVMzbTc1Y3V4bmwyV2duV05xbzdMcW04V3Vra2U5bERjcXFsSWJGeFkxckpTMWhCNU96eURjX2IyeXJyZEwzVlRyRHBndjEzcjdOUnR1MXF2VnVrcHF1MVRQTEtlYVhBRVVqX2xKekdRUjItcC1nVy16WS1XdzNYYXpUT3B0T3ZNdERxSzIwdDVobGV4ekM5dFJFSldrQndCeGgzWUZ2dXJnaGpCNGFCNDl2b3FrUkVSRVJFUkVSRVJFUlFmQ2xFUkVSRVJFUkVYWE14Yk9JUGh2eTV2TGxqc0Q5RndYTzIwdDNvS2kyVjhaZEJVeHZqa0RYRnA0SHozQzhQVDZRMHp0Tm84V0RiUFJRdDlOTTg4bzdaU1BjOHk4Y0NXVjJDLVJ4SUFkSklTNDV5U3NPLXFheTcxRFF0MHUyaXR1TlUzLTdzeUtXbHBySlUxTC1UM05ieURHTUplRzVKSUE3Z0xXVEYwbDlXVjd1ZnczZE8yNTc2cXJrYzkwMVhwaXVqYTV6Z1M0dmxraUFCUGZKYzRaSi1xeXpzWHA3YTEwTDB4dWwxRm9LN1hEY2ZjdTdXdTJRUTBWdG1xM2FZb0hUdE1zdFY4TnA0QVlEbmpzQUFBWEhCVzJiUUduckxwRFI5bjBocHlna29yVnA2amh0RkZDOXBhV3dVekJDd0RQekVBTUFEajNkNVhxRVJFUkVSRVJFUkVSRVJGQjhLVVJFUkVSRVJFUlVEd1B5cGI0X21mLTlkYWZfM2xuLUdUX0tGVkw1Zl9BSXpfQVBqVV93RFdmOFhfQURLbG5odjViX3lMbF82d193Q1BfbFhLaUlpSWlJaUlpSWlJaUl2XzJRPT0iXQ~WyJfR2p3dm5vV2VvV0pNbnFjT1d3ZndnIiwgImJpcnRoX2ZhbWlseV9uYW1lIiwgIkFOREVSU1NPTiJd~WyJfc3hPNDdqaU9jcTNxYWVMOElBdkVRIiwgImJpcnRoX2dpdmVuX25hbWUiLCAiSkFOIl0~WyItUXZUOEczU0YwNG5FZjh5M1A4ZTNBIiwgInNleCIsICIxIl0~WyJuLVpmMi1yckdVTUU4SHR5QmRFVWh3IiwgIjE2IiwgInVuc2V0Il0~WyJaWHFFVDgyaG1YUmV2Sy1XN0J2V1dBIiwgIjE4IiwgInRydWUiXQ~WyJIclE5UVNNdE9rTm1tc202UjNrUERnIiwgIjY1IiwgImZhbHNlIl0~WyJJSXFQQk5JMFd1NGlKWlluSkFZakRnIiwgImFnZV9lcXVhbF9vcl9vdmVyIiwgeyJfc2QiOiBbIjl2b2hTQUNUaVI5OURTOEgyMzFaaENCYkMySzR3SERLQXFtNlJhWFc4Q3MiLCAiUnZDYjlJa0JzclM3eHluUkVZQnEyMlpzb19UUmllLTU5ZndxZEgwaWFDayIsICJaLWdzVDMxQkRWejFndHFiT1Jqd1pBNTB0YW9kdVBacWpjZzNVczQ1VGNFIl19XQ~WyJUTmVHckxXTWprUEstbU81elRnaGpBIiwgImFnZV9pbl95ZWFycyIsIDQwXQ~WyJnNnFRSzM0SXJlSjFwNXV0UmpiR0JBIiwgImRvY3VtZW50X251bWJlciIsICIxMTExMTExMTQiXQ~WyJWUHhEMGo5OXlCdUdod3ZPTHE2VXlnIiwgImlzc3VpbmdfYXV0aG9yaXR5IiwgIlRlc3QgUElEIGlzc3VlciJd~WyJscmczQkdyaTlMUmRSdUhhYlV4dVJnIiwgImlzc3VpbmdfY291bnRyeSIsICJGQyJd~"
-
-const val mockedSdJwtPidBasicFields: String =
-    "eyJhbGciOiAiRVMyNTYiLCAidHlwIjogImRjK3NkLWp3dCIsICJ4NWMiOiBbIk1JSUMzekNDQW9XZ0F3SUJBZ0lVZjNsb2hUbURNQW1TL1lYL3E0aHFvUnlKQjU0d0NnWUlLb1pJemowRUF3SXdYREVlTUJ3R0ExVUVBd3dWVUVsRUlFbHpjM1ZsY2lCRFFTQXRJRlZVSURBeU1TMHdLd1lEVlFRS0RDUkZWVVJKSUZkaGJHeGxkQ0JTWldabGNtVnVZMlVnU1cxd2JHVnRaVzUwWVhScGIyNHhDekFKQmdOVkJBWVRBbFZVTUI0WERUSTFNRFF4TURFME16YzFNbG9YRFRJMk1EY3dOREUwTXpjMU1Wb3dVakVVTUJJR0ExVUVBd3dMVUVsRUlFUlRJQzBnTURFeExUQXJCZ05WQkFvTUpFVlZSRWtnVjJGc2JHVjBJRkpsWm1WeVpXNWpaU0JKYlhCc1pXMWxiblJoZEdsdmJqRUxNQWtHQTFVRUJoTUNWVlF3V1RBVEJnY3Foa2pPUFFJQkJnZ3Foa2pPUFFNQkJ3TkNBQVM3V0FBV3FQemUwVXMzejhwYWp5VlBXQlJtclJiQ2k1WDJzOUd2bHliUXl0d1R1bWNabmVqOUJrTGZBZ2xsb1g1dHYrTmdXZkRmZ3QvMDZzKzV0VjRsbzRJQkxUQ0NBU2t3SHdZRFZSMGpCQmd3Rm9BVVlzZVVSeWk5RDZJV0lLZWF3a21VUlBFQjA4Y3dHd1lEVlIwUkJCUXdFb0lRYVhOemRXVnlMbVYxWkdsM0xtUmxkakFXQmdOVkhTVUJBZjhFRERBS0JnZ3JnUUlDQUFBQkFqQkRCZ05WSFI4RVBEQTZNRGlnTnFBMGhqSm9kSFJ3Y3pvdkwzQnlaWEJ5YjJRdWNHdHBMbVYxWkdsM0xtUmxkaTlqY213dmNHbGtYME5CWDFWVVh6QXlMbU55YkRBZEJnTlZIUTRFRmdRVXFsL29weGtRbFl5MGxsYVRvUGJERS9teUVjRXdEZ1lEVlIwUEFRSC9CQVFEQWdlQU1GMEdBMVVkRWdSV01GU0dVbWgwZEhCek9pOHZaMmwwYUhWaUxtTnZiUzlsZFMxa2FXZHBkR0ZzTFdsa1pXNTBhWFI1TFhkaGJHeGxkQzloY21Ob2FYUmxZM1IxY21VdFlXNWtMWEpsWm1WeVpXNWpaUzFtY21GdFpYZHZjbXN3Q2dZSUtvWkl6ajBFQXdJRFNBQXdSUUloQU5KVlNEc3FUM0lrR2NLV1dnU2V1YmtET2RpNS9VRTliMUdGL1g1ZlFSRmFBaUJwNXQ2dEhoOFh3RmhQc3R6T0hNb3B2QkQvR3dtczBSQVVnbVNuNmt1OEdnPT0iXX0.eyJfc2QiOiBbIkJUd09DRlNTSW1VQkdnRVkyQTQxSnlFN08tN3ZHbmxsWmNYdzhWd2t6aEkiLCAiRHdFSktNTkllaDkteW1GMFNsaHIwSzVRNkV5ZjJMVXo5WXlJQ3l6RnFLUSIsICJFWnhPQkx0UUdsSjYtU3lXTEFObXJlQVVST3NxU1lOWW9yRUZwMG40MHhzIiwgIkY5ZWs5eU5QVDI4MlppX1J6N0pXNzZ4cThpMHJlTDdvN0ozcDMzRjdqSlkiLCAiU3c5VEF2S1pHQXVYbnJDa0Jqbkx4YU5wdUszNkJWVUlzaDhoVGZVcWVfWSIsICJYZDNHTm9oWXNORkREN2JYN3kzWks2d3RzcU85OURSYWMyckxqRnBuT3RFIiwgIlhxVk9kbzRzQjZSZlo3aFlmdnhBZW5ZSjV5dE1MWERfUlhCVzZ6SGJrNGMiLCAiWUtOWXY1MnZaUHNnM0pFcTRNU2c2X0FFQkF0dnVaT2trMXZ3UWZPam5vdyIsICJnR3pVOExNYW1CTTF4X3pEOFBOUUk4anpZVU1uaHJUa09PQU1RdVdlcVVBIl0sICJpc3MiOiAiaHR0cHM6Ly9pc3N1ZXIuZXVkaXcuZGV2IiwgImlhdCI6IDE3NDc5NTQ4MDAsICJleHAiOiAxNzU1NzMwODAwLCAidmN0IjogInVybjpldS5ldXJvcGEuZWMuZXVkaTpwaWQ6MSIsICJzdGF0dXMiOiB7ImlkZW50aWZpZXJfbGlzdCI6IHsiaWQiOiAiNzgwMyIsICJ1cmkiOiAiaHR0cHM6Ly9pc3N1ZXIuZXVkaXcuZGV2L2lkZW50aWZpZXJfbGlzdC9GQy9ldS5ldXJvcGEuZWMuZXVkaS5waWQuMS85OTg3ZGNhNS1mMzAxLTRmNzUtYTllZi1kNjJiMDk1YTBlZjQifSwgInN0YXR1c19saXN0IjogeyJpZHgiOiA3ODAzLCAidXJpIjogImh0dHBzOi8vaXNzdWVyLmV1ZGl3LmRldi90b2tlbl9zdGF0dXNfbGlzdC9GQy9ldS5ldXJvcGEuZWMuZXVkaS5waWQuMS85OTg3ZGNhNS1mMzAxLTRmNzUtYTllZi1kNjJiMDk1YTBlZjQifX0sICJfc2RfYWxnIjogInNoYS0yNTYiLCAiY25mIjogeyJqd2siOiB7Imt0eSI6ICJFQyIsICJjcnYiOiAiUC0yNTYiLCAieCI6ICJGOWdpRy02Wmh6WV80Rm13WGtqX191UzBLOTZhMzdURDJxNjVQelAtY3M0IiwgInkiOiAiMHd0Y19IdTc1QnI4XzZabnctcHEtSUdlUjhpVkdFOUdJVElHdFBQRmRMayJ9fX0.2Zuy9m9Y9qj13BID7U8er--byhTta5m7xxuSv1i4udn7ydMX8Zb4uKBCutzQBe-1b7D5_qurPbYTJbCRQVqT1A~WyJRT0lhYmgtV2tlSU1kZVlCdFhvUEtBIiwgImZhbWlseV9uYW1lIiwgIkFOREVSU1NPTiJd~WyIyTHlJcjJWeWFKSFRsdzlRU3cwd3VBIiwgImdpdmVuX25hbWUiLCAiSkFOIl0~WyJyQnVOVDBwbTl3LUJGSUYwWkpRbkJnIiwgImJpcnRoX2RhdGUiLCAiMTk4NS0wMy0zMCJd~WyJzdVhSSl9JLXVjbk1vVVctN1Z4RmpBIiwgImxvY2FsaXR5IiwgIktBVFJJTkVIT0xNIl0~WyJQUmJKQlZ2eWxaV2dCTndVajRYUmpRIiwgInBsYWNlX29mX2JpcnRoIiwgeyJfc2QiOiBbIlJ3WEZMOTJicFU0YmFlNmhDTGMyVFlBc2ZmaVc0U1ZCQ2tSYUtseHJtN3ciXX1d~WyJWRWpnOW1ROGV2d3VtRGE3bTc1WVB3IiwgIm5hdGlvbmFsaXRpZXMiLCBbIlNFIl1d~WyJvbTVpYUxVdzRNblZSelQ2Nk53bmtnIiwgIjE4IiwgInRydWUiXQ~WyJPUDA5OU5ZN29LcFM1R1pQRDZrU19RIiwgIjY1IiwgInVuc2V0Il0~WyJVYTA0TFJOc05DUjF6SU96VGRTaTlBIiwgImFnZV9lcXVhbF9vcl9vdmVyIiwgeyJfc2QiOiBbIi05cm16RVJJWHl5X0trQWdBbC1vMF82RGJad05lWjZ0Y3FLRzl1V0RhX0kiLCAicDBQNXROU1VxYXdQU2p4dVdmQ3paZmNaNENfYVFTV0V3TmhjT28zZ3ZkNCJdfV0~WyJXWEdxdm9zR0RuOWNaWm11Y2FSdFFBIiwgImFnZV9iaXJ0aF95ZWFyIiwgMTk4NV0~WyI3QTQtNkZHaURhRThfQ1laTEw0OERnIiwgImlzc3VpbmdfYXV0aG9yaXR5IiwgIlRlc3QgUElEIGlzc3VlciJd~WyJVU0JRX2ZhaUl1X0pkQnl3dnVZd1dnIiwgImlzc3VpbmdfY291bnRyeSIsICJGQyJd~"
-
-val mockedPidBasicFields: Map<String, ByteArray> = mapOf(
-    "family_name" to byteArrayOf(105, 65, 78, 68, 69, 82, 83, 83, 79, 78),
-    "given_name" to byteArrayOf(99, 74, 65, 78),
-    "age_over_18" to byteArrayOf(-11),
-    "age_over_65" to byteArrayOf(-12),
-    "age_birth_year" to byteArrayOf(25, 7, -63),
-    "birth_city" to byteArrayOf(107, 75, 65, 84, 82, 73, 78, 69, 72, 79, 76, 77),
-    "gender" to byteArrayOf(1),
-    "expiry_date" to byteArrayOf(
-        -64,
-        116,
-        50,
-        48,
-        53,
-        48,
-        45,
-        48,
-        51,
-        45,
-        51,
-        48,
-        84,
-        48,
-        48,
-        58,
-        48,
-        48,
-        58,
-        48,
-        48,
-        90
-    ),
-)
-
-val mockedMdlBasicFields: Map<String, ByteArray> = mapOf(
-    "family_name" to byteArrayOf(105, 65, 78, 68, 69, 82, 83, 83, 79, 78),
-    "given_name" to byteArrayOf(99, 74, 65, 78),
-    "birth_place" to byteArrayOf(102, 83, 87, 69, 68, 69, 78),
-    "expiry_date" to byteArrayOf(
-        -64,
-        116,
-        50,
-        48,
-        53,
-        48,
-        45,
-        48,
-        51,
-        45,
-        51,
-        48,
-        84,
-        48,
-        48,
-        58,
-        48,
-        48,
-        58,
-        48,
-        48,
-        90
-    ),
-    "portrait" to byteArrayOf(98, 83, 69),
-
-    "signature_usual_mark" to byteArrayOf(98, 83, 69),
-    "sex" to byteArrayOf(1),
-)
-
-const val mockedPidDocType = "eu.europa.ec.eudi.pid.1"
-const val mockedPidNameSpace = "eu.europa.ec.eudi.pid.1"
-val mockedMsoMdocPidFormat = MsoMdocFormat(mockedPidDocType)
-val mockedSdJwtPidFormat = SdJwtVcFormat("urn:eudi:pid:1")
-const val mockedMdlDocType = "org.iso.18013.5.1.mDL"
-const val mockedMdlNameSpace = "org.iso.18013.5.1"
-val mockedMsoMdocMdlFormat = MsoMdocFormat(mockedMdlDocType)
-const val mockedDocDisplayLogo = "https://examplestate.com/public/pid.png"
-const val mockedIssuerLogo = "https://issuer.eudiw.dev/ic-logo.png"
-const val mockedIssuerIdentifier = "https://issuer.eudiw.dev"
-
-fun createMockedNamespaceData(
-    documentNamespace: String,
-    nameSpacedData: Map<String, ByteArray>,
-): NameSpacedData {
-    val builder = NameSpacedData.Builder()
-    nameSpacedData.forEach {
-        builder.putEntry(documentNamespace, it.key, it.value)
-    }
-    return builder.build()
-}
-
-suspend fun mockIssuedDocument(
-    format: DocumentFormat,
-    id: String,
-    name: String,
-    isCertified: Boolean,
-    createdAt: Instant,
-    issuedAt: Instant,
-    validFrom: Result<Instant>,
-    validUntil: Result<Instant>,
-    data: DocumentData,
-    issuerMetadata: IssuerMetadata? = null,
-    availableCredentials: Int,
-    totalCredentials: Int,
-): IssuedDocument {
-    val doc = mock<IssuedDocument>()
-
-    with(doc) {
-        whenever(this.format)
-            .thenReturn(format)
-
-        whenever(this.id)
-            .thenReturn(id)
-
-        whenever(this.name)
-            .thenReturn(name)
-
-        whenever(this.isCertified())
-            .thenReturn(isCertified)
-
-        whenever(this.createdAt)
-            .thenReturn(createdAt)
-
-        whenever(this.issuedAt)
-            .thenReturn(issuedAt)
-
-        whenever(this.getValidFrom())
-            .thenReturn(validFrom)
-
-        whenever(this.getValidUntil())
-            .thenReturn(validUntil)
-
-        whenever(this.data)
-            .thenReturn(data)
-
-        whenever(this.issuerMetadata)
-            .thenReturn(issuerMetadata)
-
-        whenever(this.credentialsCount())
-            .thenReturn(availableCredentials)
-
-        whenever(this.initialCredentialsCount())
-            .thenReturn(totalCredentials)
-    }
-
-    return doc
-}
-
-suspend fun getMockedFullPid() = mockIssuedDocument(
-    format = mockedMsoMdocPidFormat,
-    id = mockedPidId,
-    name = mockedPidDocName,
-    isCertified = false,
-    createdAt = Instant.parse(mockedDocumentCreationDate),
-    issuedAt = Instant.parse(mockedDocumentCreationDate),
-    validFrom = Result.success(Instant.now()),
-    validUntil = Result.success(Instant.parse(mockedDocumentValidUntilDate)),
-    data = MsoMdocData(
-        format = mockedMsoMdocPidFormat,
-        issuerMetadata = null,
-        nameSpacedData = createMockedNamespaceData(
-            mockedPidNameSpace,
-            mockedPidFields
-        )
-    ),
-    availableCredentials = mockedDocumentAvailableCredentials,
-    totalCredentials = mockedDocumentTotalCredentials,
-)
-
-suspend fun getMockedSdJwtFullPid() = mockIssuedDocument(
-    format = mockedSdJwtPidFormat,
-    id = mockedSdJwtPidId,
-    name = mockedPidDocName,
-    isCertified = false,
-    createdAt = Instant.parse(mockedDocumentCreationDate),
-    issuedAt = Instant.parse(mockedDocumentCreationDate),
-    validFrom = Result.success(Instant.now()),
-    validUntil = Result.success(Instant.parse(mockedDocumentValidUntilDate)),
-    data = SdJwtVcData(
-        format = mockedSdJwtPidFormat,
-        issuerMetadata = null,
-        sdJwtVc = mockedSdJwtFullPidFields,
-    ),
-    availableCredentials = mockedDocumentAvailableCredentials,
-    totalCredentials = mockedDocumentTotalCredentials,
-)
-
-suspend fun getMockedMainPid() = getMockedFullPid()
-
-suspend fun IssuedDocument.copy(
-    format: DocumentFormat? = null,
-    id: String? = null,
-    name: String? = null,
-    isCertified: Boolean? = null,
-    createdAt: Instant? = null,
-    issuedAt: Instant? = null,
-    validFrom: Result<Instant>? = null,
-    validUntil: Result<Instant>? = null,
-    data: DocumentData? = null,
-    issuerMetadata: IssuerMetadata? = null,
-    availableCredentials: Int? = null,
-    totalCredentials: Int? = null,
-): IssuedDocument {
-    if (format != null) {
-        whenever(this.format)
-            .thenReturn(format)
-    }
-
-    if (id != null) {
-        whenever(this.id)
-            .thenReturn(id)
-    }
-
-    if (name != null) {
-        whenever(this.name)
-            .thenReturn(name)
-    }
-
-    if (isCertified != null) {
-        whenever(this.isCertified())
-            .thenReturn(isCertified)
-    }
-
-    if (createdAt != null) {
-        whenever(this.createdAt)
-            .thenReturn(createdAt)
-    }
-
-    if (issuedAt != null) {
-        whenever(this.issuedAt)
-            .thenReturn(issuedAt)
-    }
-
-    if (validFrom != null) {
-        whenever(this.getValidFrom())
-            .thenReturn(validFrom)
-    }
-
-    if (validUntil != null) {
-        whenever(this.getValidUntil())
-            .thenReturn(validUntil)
-    }
-
-    if (data != null) {
-        whenever(this.data)
-            .thenReturn(data)
-    }
-
-    if (issuerMetadata != null) {
-        whenever(this.issuerMetadata)
-            .thenReturn(issuerMetadata)
-    }
-
-    if (availableCredentials != null) {
-        whenever(this.credentialsCount())
-            .thenReturn(availableCredentials)
-    }
-
-    if (totalCredentials != null) {
-        whenever(this.initialCredentialsCount())
-            .thenReturn(totalCredentials)
-    }
-
-    return this
-}
-
-suspend fun getMockedPidWithBasicFields() = getMockedFullPid().copy(
-    data = MsoMdocData(
-        format = mockedMsoMdocPidFormat,
-        issuerMetadata = null,
-        nameSpacedData = createMockedNamespaceData(
-            mockedPidNameSpace,
-            mockedPidBasicFields
-        )
-    )
-)
-
-val mockedDocumentMetadata = IssuerMetadata(
-    documentConfigurationIdentifier = mockedPidDocType,
-    display = listOf(
-        IssuerMetadata.Display(
-            name = mockedPidDocName,
-            logo = IssuerMetadata.Logo(
-                uri = URI.create(mockedDocDisplayLogo)
-            )
-        )
-    ),
-    claims = emptyList(),
-    credentialIssuerIdentifier = mockedIssuerIdentifier,
-    issuerDisplay = listOf(
-        IssuerMetadata.IssuerDisplay(
-            name = "EUDIW Issuer",
-            logo = IssuerMetadata.Logo(
-                uri = URI.create(mockedIssuerLogo)
-            )
-        )
-    )
-)
-
-suspend fun getMockedPidWithBasicFieldsAndMetadata() = getMockedFullPid().copy(
-    issuerMetadata = mockedDocumentMetadata,
-    data = MsoMdocData(
-        format = mockedMsoMdocPidFormat,
-        issuerMetadata = mockedDocumentMetadata,
-        nameSpacedData = createMockedNamespaceData(
-            mockedPidNameSpace,
-            mockedPidBasicFields
-        )
-    )
-)
-
-suspend fun getMockedSdJwtPidWithBasicFields() = getMockedSdJwtFullPid().copy(
-    data = SdJwtVcData(
-        format = mockedSdJwtPidFormat,
-        issuerMetadata = null,
-        sdJwtVc = mockedSdJwtPidBasicFields,
-    )
-)
-
-suspend fun getMockedOldestPidWithBasicFields() = getMockedPidWithBasicFields().copy(
-    id = mockedOldestPidId,
-    createdAt = Instant.parse(mockedOldestDocumentCreationDate)
-)
-
-suspend fun getMockedEmptyPid() = getMockedFullPid().copy(
-    data = MsoMdocData(
-        format = mockedMsoMdocPidFormat,
-        issuerMetadata = null,
-        nameSpacedData = createMockedNamespaceData(
-            mockedPidNameSpace,
-            emptyMap()
-        )
-    )
-)
-
-suspend fun getMockedFullMdl() = mockIssuedDocument(
-    format = mockedMsoMdocMdlFormat,
-    id = mockedMdlId,
-    name = mockedMdlDocName,
-    isCertified = false,
-    createdAt = Instant.parse(mockedDocumentCreationDate),
-    issuedAt = Instant.parse(mockedDocumentCreationDate),
-    validFrom = Result.success(Instant.now()),
-    validUntil = Result.success(Instant.parse(mockedDocumentValidUntilDate)),
-    data = MsoMdocData(
-        format = mockedMsoMdocMdlFormat,
-        issuerMetadata = null,
-        nameSpacedData = createMockedNamespaceData(
-            mockedMdlNameSpace,
-            mockedMdlFields
-        )
-    ),
-    availableCredentials = mockedDocumentAvailableCredentials,
-    totalCredentials = mockedDocumentTotalCredentials,
-)
-
-suspend fun getMockedMdlWithBasicFields() = getMockedFullMdl().copy(
-    data = MsoMdocData(
-        format = mockedMsoMdocMdlFormat,
-        issuerMetadata = null,
-        nameSpacedData = createMockedNamespaceData(
-            mockedMdlNameSpace,
-            mockedMdlBasicFields
-        )
-    )
-)
-
-suspend fun getMockedMdlWithNoExpirationDate(): IssuedDocument = getMockedFullMdl().copy(
-    data = MsoMdocData(
-        format = mockedMsoMdocMdlFormat,
-        issuerMetadata = null,
-        nameSpacedData = createMockedNamespaceData(
-            mockedMdlNameSpace,
-            mockedMdlFields
-                .minus("expiry_date"),
-        )
-    )
-)
-
-suspend fun getMockedMdlWithNoUserNameAndNoUserImage(): IssuedDocument = getMockedFullMdl().copy(
-    data = MsoMdocData(
-        format = mockedMsoMdocMdlFormat,
-        issuerMetadata = null,
-        nameSpacedData = createMockedNamespaceData(
-            mockedMdlNameSpace,
-            mockedMdlFields
-                .minus("given_name")
-                .minus("portrait")
-        )
-    )
-)
-
-suspend fun getMockedFullDocuments(): List<IssuedDocument> = listOf(
-    getMockedFullPid(), getMockedFullMdl()
-)
